+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Verion 1.0
 -Basic Model
 -Basic Input Validation; on;y checks if the entered integers are between 0 or 2
--Game Logic Runs
+-playable if input isnt invalid
+
+### v1.0 bugs
+-if there's an invalid move, the game passes the turn again to the next player
+-too many boards